server/requirements.txt
fastapi==0.115.12
python-multipart>=0.0.9
uvicorn[standard]==0.34.2
aiosqlite==0.21.0
openpyxl==3.1.5
pydantic-settings==2.9.1
python-dotenv==1.1.0
lxml==5.4.0
httpx==0.28.1
watchdog==6.0.0
Pillow==12.1.1
gspread==6.2.1
pywin32; sys_platform == "win32"
pytest==8.3.5
pytest-asyncio==0.25.3
